+++ src/tests/RuleSets/Rules/TypeScript/MemberOrdering/SourceCodeCreator.ts
@@ -109,6 +109,20 @@ export class SourceCodeCreator
     }
 
     /**
+     * Generates test-code for class-members.
+     *
+     * @param members
+     * The class-members to create.
+     *
+     * @returns
+     * The code of the specified `members`.
+     */
+    public static TestMembers(members: MemberDefinition[]): string
+    {
+        return this.GetContainerCode(ContainerType.Class, ...members);
+    }
+
+    /**
      * Gets code for a container.
      *
      * @param containerType
@@ -245,7 +259,7 @@ export class SourceCodeCreator
             case MemberType.Constructor:
                 code = isClass ? "constructor() { }" : "new ();";
                 break;
-            case MemberType.Signature:
+            case MemberType.Indexer:
                 code = "[key: string]: any;";
                 break;
             case MemberType.Method: