+++ src/tests/RuleSets/Rules/TypeScript/NoDynamicDelete.ts
@@ -6,42 +6,19 @@ export let NoDynamicDelete = new RuleSuite(
     "@typescript-eslint/no-dynamic-delete",
     [
         {
-            Description: "Checking whether the rule works as expected…",
+            Description: "Checking whether dynamic deletes are allowed…",
             RuleSet: RuleSet.Recommended,
             ScriptKind: ScriptKind.Scripts,
             CodeSnippets: [
-                {
-                    Valid: false,
-                    Snippets: [
-                        "delete container[variable];",
-                        "delete container[variable.toUpperCase()];",
-                        "delete container['abc'];"
-                    ]
-                },
                 {
                     Valid: true,
                     Snippets: [
                         "delete container[1];",
                         "delete container['-prop-that-must-be-quoted'];",
-                        "delete container.abc;"
-                    ]
-                }
-            ]
-        },
-        {
-            Description: "Checking whether this rule is deactivated for the weak rule-set…",
-            RuleSet: RuleSet.Weak,
-            ScriptKind: ScriptKind.Scripts,
-            CodeSnippets: [
-                {
-                    Valid: true,
-                    Snippets: [
+                        "delete container.abc;",
                         "delete container[variable];",
                         "delete container[variable.toUpperCase()];",
-                        "delete container['abc'];",
-                        "delete container[1];",
-                        "delete container['-quoted-prop'];",
-                        "delete container.abc;"
+                        "delete container['abc'];"
                     ]
                 }
             ]